requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client==1.7.0
anthropic==0.21.3
python-multipart==0.0.6
python-dotenv==1.0.0
asyncio==3.4.3
requests==2.31.0